Rename a11y-result -> content checker files
