+++ client/scss/components/_content-checker.scss
