+++ client/scss/core.scss
@@ -137,7 +137,7 @@ These are classes for components.
 @use 'components/preview-panel';
 @use 'components/preview-error';
 @use 'components/form-side';
-@use 'components/a11y-result';
+@use 'components/content-checker';
 @use 'components/userbar';
 @use 'components/breadcrumbs';
 @use 'components/pill';
