+++ client/src/includes/contentChecker.ts
